chaging function name from check_jwt to login_required
