+++ students/kimjy/westagram/get_token.sh
@@ -1,0 +1,1 @@
+http POST 127.0.0.1:8080/users/signin email="sseunn@naver.com" password="1234qwer@"
